单据扫描界面修改价格添加权限限制

+++ app/src/main/java/com/eshop/jinxiaocun/lingshou/bean/GetOptAuthResult.java
@@ -10,7 +10,7 @@ public class GetOptAuthResult extends BaseResult {
     private double limitdiscount;//:80 //折扣最小额度
     private double savediscount;// : 60//折扣最大额度
 
-    public double getIsgrant() {
+    public int getIsgrant() {
         return isgrant;
     }
